+++ Editor/Unity.Platforms.Build.DotsRuntime/IDotsRuntimeBuildModifier.cs
@@ -1,9 +1,0 @@
-using Newtonsoft.Json.Linq;
-
-namespace Unity.Platforms.Build.DotsRuntime
-{
-    public interface IDotsRuntimeBuildModifier : IBuildComponent
-    {
-        void Modify(JObject settingsJObject);
-    }
-}