+++ Editor/Unity.Platforms.Build/IBuildArtifacts.cs
@@ -1,9 +1,0 @@
-﻿namespace Unity.Platforms.Build
-{
-    /// <summary>
-    /// Interface for information about a build.
-    /// </summary>
-    public interface IBuildArtifact
-    {
-    }
-}