+++ Editor/Unity.Platforms.Build/IRunInstance.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Unity.Platforms.Build
-{
-    public interface IRunInstance : IDisposable
-    {
-        bool IsRunning { get; }
-    }
-}